Apply suggested changes. Use dataclasses for entries in Node Process and History tables. Move start_node function and _node class initialization to NodePM. Add and modify unit tests.
